backend/requirements.txt
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Deprecated==1.3.1
dnspython==2.8.0
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==3.11.0
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==26.0
Pygments==2.20.0
pymongo==4.7.0
python-dotenv==1.2.1
requests==2.32.5
rich==13.9.4
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wrapt==2.1.2
